+++ js/Helper.js
@@ -41,3 +41,16 @@ export function generateUUID() {
     return (c === "x" ? r : (r & 0x3) | 0x8).toString(16);
   });
 }
+
+/* copied from https://www.w3resource.com/javascript-exercises/javascript-date-exercise-24.php */
+export function getWeekNumber(dt) {
+  let tdt = new Date(dt.valueOf());
+  let dayn = (dt.getDay() + 6) % 7;
+  tdt.setDate(tdt.getDate() - dayn + 3);
+  let firstThursday = tdt.valueOf();
+  tdt.setMonth(0, 1);
+  if (tdt.getDay() !== 4) {
+    tdt.setMonth(0, 1 + ((4 - tdt.getDay() + 7) % 7));
+  }
+  return 1 + Math.ceil((firstThursday - tdt) / 604800000);
+}
