+++ js/screens/NewEntry.js
@@ -20,6 +20,7 @@ import { Ionicons } from "@expo/vector-icons";
 import { addTemplate } from "../redux/actions/User";
 import { addFinanceItem } from "../redux/actions/Finance";
 import firebase from "../Firebase";
+import { generateUUID, selectablePaymentMethods, uploadImage } from "../Helper";
 import AppSafeAreaView from "../components/AppSafeAreaView";
 import { ObjectItemPicker, StringItemPicker } from "../components/ItemPicker";
 import NewEntry_Template from "../components/NewEntry_Template";
@@ -359,13 +360,6 @@ function NewEntry(props) {
   );
 }
 
-const selectablePaymentMethods = [
-  { label: "Barzahlung", value: "cash" },
-  { label: "EC-Karte", value: "debit-card" },
-  { label: "Kreditkarte", value: "credit-card" },
-  { label: "PayPal", value: "paypal" },
-];
-
 const styles = StyleSheet.create({
   container: {
     flex: 1,
@@ -418,39 +412,3 @@ const mapDispatchToProps = (dispatch) =>
   bindActionCreators({ addFinanceItem, addTemplate }, dispatch);
 
 export default connect(mapStateToProps, mapDispatchToProps)(NewEntry);
-
-/* https://stackoverflow.com/questions/105034/how-to-create-a-guid-uuid */
-function generateUUID() {
-  // Public Domain/MIT
-  var d = new Date().getTime(); //Timestamp
-  var d2 = (performance && performance.now && performance.now() * 1000) || 0; //Time in microseconds since page-load or 0 if unsupported
-  return "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g, function (c) {
-    var r = Math.random() * 16; //random number between 0 and 16
-    if (d > 0) {
-      //Use timestamp until depleted
-      r = (d + r) % 16 | 0;
-      d = Math.floor(d / 16);
-    } else {
-      //Use microseconds since page-load if supported
-      r = (d2 + r) % 16 | 0;
-      d2 = Math.floor(d2 / 16);
-    }
-    return (c === "x" ? r : (r & 0x3) | 0x8).toString(16);
-  });
-}
-
-async function uploadImage(uri) {
-  const response = await fetch(uri);
-  const imageBlob = await response.blob();
-
-  const metadata = {
-    contentType: "image/jpeg",
-  };
-  const name = generateUUID() + "-media.jpg";
-  const imageRef = firebase.storage.ref().child("images/" + name);
-
-  const snapshot = await imageRef.put(imageBlob, metadata);
-  const downloadURL = await snapshot.ref.getDownloadURL();
-
-  return downloadURL;
-}
